Extract SummerBoltProvider
